# src/openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/images/generations": {
      "post": {
        "operationId": "generateImage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageGenerationRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageGenerationResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "ComfyUI unavailable or generation failed"
          }
        },
        "summary": "Generate an image using the zib-zit-moody ComfyUI workflow",
        "tags": [
          "images"
        ]
      }
    },
...
  },
  "info": {
    "title": "AI Proxy API",
    "description": "OpenAI-compatible proxy for llama.cpp with context compression, retry, and tool-call buffering",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "ImageGenerationRequestDto": {
        "type": "object",
        "properties": {
          "prompt": {
            "type": "string",
            "description": "Text prompt for image generation"
          },
          "negativePrompt": {
            "type": "string",
            "description": "Proxy extension: negative prompt for things to avoid in the image"
          },
          "model": {
            "type": "string",
            "description": "Model name (currently always maps to zib-zit-moody workflow)"
          },
          "n": {
            "type": "number",
            "description": "Number of images to generate (currently only 1 is supported)",
            "example": 1
          },
          "size": {
            "type": "string",
            "description": "Image size as WxH string",
            "example": "1024x1024"
          },
          "response_format": {
            "type": "string",
            "description": "Response format",
            "example": "b64_json",
            "enum": [
              "url",
              "b64_json"
            ]
          },
          "quality": {
            "type": "string",
            "description": "Image quality hint (ignored, passed through)",
            "example": "standard"
          },
          "style": {
            "type": "string",
            "description": "Image style hint (ignored, passed through)",
            "example": "vivid"
          }
        },
        "required": [
          "prompt"
        ]
      },
      "ImageDataDto": {
        "type": "object",
        "properties": {
          "b64_json": {
            "type": "string",
            "description": "Base64-encoded image data"
          },
          "url": {
            "type": "string",
            "description": "URL of the generated image"
          },
          "revised_prompt": {
            "type": "string",
            "description": "The revised prompt used for generation"
          }
        }
      },
      "ImageGenerationResponseDto": {
        "type": "object",
        "properties": {
          "created": {
            "type": "number",
            "description": "Unix timestamp of when the request was made"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImageDataDto"
            }
          }
        },
        "required": [
          "created",
          "data"
        ]
      }
    }
  }
}